+++ src/components/MetadataCard/MetadataCard.cy.tsx
@@ -2,10 +2,17 @@ import React from 'react';
 import { CssBaseline, ThemeProvider } from '@mui/material';
 import { THEME_DARK, THEME_LIGHT } from '@ska-telescope/ska-gui-components';
 import theme from '@services/theme/theme';
-import mockData from '@services/Mocking/mockMetaData';
 import MetadataCard from '@components/MetadataCard/MetadataCard';
 
 const THEME = [THEME_DARK, THEME_LIGHT];
+
+const emptyDataProduct = {
+  execution_block: '',
+  relativePathName: '',
+  metaDataFile: '',
+  uid: '',
+  metadata_store_name: ''
+};
 
 describe('<MetadataCard />', () => {
   for (const theTheme of THEME) {
@@ -13,10 +20,9 @@ describe('<MetadataCard />', () => {
       cy.mount(
         <ThemeProvider theme={theme(theTheme)}>
           <CssBaseline />
-          <MetadataCard metaData={mockData} />
+          <MetadataCard {...emptyDataProduct} />
         </ThemeProvider>
       );
-      cy.findByTestId('MetadataDataTree').should('be.visible');
     });
   }
 });